[FIX] Filter: Improved compatibility for sqMass filtering
